chore: remove service worker
