+++ package.json
@@ -60,7 +60,6 @@
     "typescript": "^5.0.4",
     "vite": "^3.2.3",
     "vite-plugin-node-polyfills": "^0.7.0",
-    "vite-plugin-pwa": "^0.16.6",
     "workbox-core": "^6.5.4",
     "workbox-precaching": "^6.5.4",
     "workbox-routing": "^6.5.4",
